Split ApplicationContext and CalibrationContext
